+++ src/pages/MainPage/index.jsx
@@ -616,6 +616,12 @@ const Disclosure = styled.details`
     list-style: none;
     user-select: none;
     width: fit-content;
+    border-radius: ${props => props.theme.radius.sm};
+  }
+
+  summary:focus-visible {
+    outline: 2px solid ${props => props.theme.color.accent};
+    outline-offset: 2px;
   }
 
   summary::-webkit-details-marker {